test: fix cascade in 12 test files to match new Int64/typealias API
After the library refactor (Primitive widened + Phases A-D + F), test
sites that hard-coded `Int` for fields now typed as Int64 / EpochNumber
/ SlotNumber / Coin needed type annotations or casts.

Mechanical changes:
- `let epoch: Int = 540` → `let epoch: EpochNumber = 540` in
  PoolRetirementTests.
- `let slot: Int` → `let slot: SlotNumber` in After/BeforeScriptTests.
- `OrderedDictionary<String, Int>` → `<String, Int64>` test data in
  CostModelsTests.
- `Int.max` → `Int64.max` for the overflow-edge tests in
  ExecutionUnitsTests and TransactionTests.
- `let coin: Int64 = 1_000_000` in ValueTests.
- `let ttl: SlotNumber`, `validityStart: SlotNumber` in
  TransactionBodyTests.
- `Int64(...)`/`UInt64(...)` wraps at Primitive boundary in
  ListOrSetTests' MockCBORSerializable, OperationalCertTests,
  UTxOTests.
- `Int64` typed locals for stakeRegistrationDeposit / rewardBalance in
  StakeAddressInfoTests.

All assertions preserved verbatim — only annotations/casts added.

**Result: 734/734 tests pass in 7.3 seconds on macOS.** The watchOS
Int64 refactor is functionally complete on the macOS toolchain.

+++ Tests/SwiftCardanoCoreTests/PlutusTests/CostModelsTests.swift
@@ -9,7 +9,7 @@ import Testing
     @Test("Initialize with all Plutus versions")
     func testInitFromDictionary() async throws {
         // Test initialization with all Plutus versions
-        let testData: [Int: OrderedDictionary<String, Int>] = [
+        let testData: [Int: OrderedDictionary<String, Int64>] = [
             0: OrderedDictionary(
                 uniqueKeysWithValues: ["addInteger-cpu-arguments-intercept": 100788]
             ),
@@ -34,7 +34,7 @@ import Testing
     @Test("Initialize with missing Plutus versions")
     func testInitWithMissingVersions() async throws {
         // Test initialization with missing Plutus versions
-        let testData: [Int: OrderedDictionary<String, Int>] = [
+        let testData: [Int: OrderedDictionary<String, Int64>] = [
             0: OrderedDictionary(
                 uniqueKeysWithValues: ["addInteger-cpu-arguments-intercept": 100788]
             )
